Refactor pooling operation
- Pooling operation moved out of Linen
- Pooling documentation duplicated in nnx
