+++ docs_nnx/api_reference/flax.nnx/nn/index.rst
@@ -14,6 +14,7 @@ See the `NNX page <https://flax.readthedocs.io/en/latest/nnx/index.html>`__ for 
   linear
   lora
   normalization
+  pooling
   recurrent
   stochastic
